feat: add threadIds logic

+++ src/social-service/src/main/java/org/palette/easelsocialservice/dto/response/ThreadResponse.java
@@ -3,7 +3,7 @@ package org.palette.easelsocialservice.dto.response;
 import java.util.List;
 
 public record ThreadResponse(
-   Long theadId,
+   Integer theadId,
    List<PaintResponse> paints
 
 ) {
